Add cards using AnkiConnect
Updates format script to convert notes in CSV form to a JSON request
to a minimal local server created by the AnkiConnect add-on which adds
the notes to Anki. The add-to-anki applescript and script to detect
the card types are now unnecessary as the adding is done directly from
the add.py script.